Made Aweful UI
created a really bad AI for the game but I made It at 1 am and without any knowledge of tkinter so it'll do. The UI isn't implimented yet and still needs a lot of work to impliment. I uploaded so that magnus can make changes to it and fix my aweful tkinter if he wants.