Fixed buttons and slider for skillmenu. read description
I fixed the button and slider so that they now close but they still need implimentation in all the functions. The way I did it is a bit stupid and overly complicated and fragile so if you want me to impliment it that's fine or if you want you can but just make a backup first. thanks